Move DB access layer to `services`. fix typo
