+++ src/routes/apiDirectory.js
@@ -6,7 +6,7 @@ import authMiddleware, { login } from '../middleware/authMiddleware';
 import handleErrors from '../lib/apiErrorHandler';
 import adminRouter from './user';
 import activityReportsRouter from './activityReports';
-import { userById } from '../lib/users';
+import { userById } from '../services/users';
 
 export const loginPath = '/login';
